chore: move onboarding flow and dialog under same location
