+++ frontend/src/component/onboarding/dialog/CodeRenderer.tsx
@@ -1,7 +1,7 @@
 import type { CodeComponent } from 'react-markdown/lib/ast-to-react';
 import type { FC } from 'react';
 import copy from 'copy-to-clipboard';
-import useToast from '../../hooks/useToast';
+import useToast from 'hooks/useToast';
 import { IconButton, styled, Tooltip } from '@mui/material';
 import CopyIcon from '@mui/icons-material/FileCopy';
 import type { SdkName } from './sharedTypes';
