+++ frontend/src/component/onboarding/dialog/ConnectSdkDialog.tsx
