+++ frontend/src/component/onboarding/dialog/SharedComponents.tsx
