+++ frontend/src/component/onboarding/dialog/Stepper.tsx
