+++ frontend/src/component/onboarding/dialog/sharedTypes.ts
@@ -1,18 +1,18 @@
-import node from '../../assets/icons/sdks/Logo-node.svg';
-import go from '../../assets/icons/sdks/Logo-go.svg';
-import ruby from '../../assets/icons/sdks/Logo-ruby.svg';
-import php from '../../assets/icons/sdks/Logo-php.svg';
-import rust from '../../assets/icons/sdks/Logo-rust.svg';
-import dotnet from '../../assets/icons/sdks/Logo-net.svg';
-import java from '../../assets/icons/sdks/Logo-java.svg';
-import python from '../../assets/icons/sdks/Logo-python.svg';
-import javascript from '../../assets/icons/sdks/Logo-javascript.svg';
-import react from '../../assets/icons/sdks/Logo-react.svg';
-import vue from '../../assets/icons/sdks/Logo-vue.svg';
-import svelte from '../../assets/icons/sdks/Logo-svelte.svg';
-import swift from '../../assets/icons/sdks/Logo-swift.svg';
-import android from '../../assets/icons/sdks/Logo-android.svg';
-import flutter from '../../assets/icons/sdks/Logo-flutter.svg';
+import node from '../../../assets/icons/sdks/Logo-node.svg';
+import go from '../../../assets/icons/sdks/Logo-go.svg';
+import ruby from '../../../assets/icons/sdks/Logo-ruby.svg';
+import php from '../../../assets/icons/sdks/Logo-php.svg';
+import rust from '../../../assets/icons/sdks/Logo-rust.svg';
+import dotnet from '../../../assets/icons/sdks/Logo-net.svg';
+import java from '../../../assets/icons/sdks/Logo-java.svg';
+import python from '../../../assets/icons/sdks/Logo-python.svg';
+import javascript from '../../../assets/icons/sdks/Logo-javascript.svg';
+import react from '../../../assets/icons/sdks/Logo-react.svg';
+import vue from '../../../assets/icons/sdks/Logo-vue.svg';
+import svelte from '../../../assets/icons/sdks/Logo-svelte.svg';
+import swift from '../../../assets/icons/sdks/Logo-swift.svg';
+import android from '../../../assets/icons/sdks/Logo-android.svg';
+import flutter from '../../../assets/icons/sdks/Logo-flutter.svg';
 
 export type SdkType = 'client' | 'frontend';
 export type Sdk = { name: SdkName; type: SdkType };
