+++ frontend/src/component/onboarding/dialog/snippets/android.md
