+++ frontend/src/component/onboarding/dialog/snippets/go.md
