+++ frontend/src/component/onboarding/dialog/snippets/java.md
