+++ frontend/src/component/onboarding/dialog/snippets/javascript.md
