+++ frontend/src/component/onboarding/dialog/snippets/python.md
