+++ frontend/src/component/onboarding/dialog/snippets/react.md
