+++ frontend/src/component/onboarding/dialog/snippets/rust.md
